### 2021-10-26 #3853 @whedon
[PRE REVIEW]: Torque-limited simple pendulum: A toolkit for getting familiar with control algorithms in underactuated robotics
**Submitting author:** @fwiebe (<a href="http://orcid.org/0000-0002-2195-9697">Felix Wiebe</a>)
**Repository:** <a href="https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum" target ="_blank">https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum</a>
**Version:** v1.0.0
