### 2021-10-27 9b80aae
ilqr implementation with sympy
