+++ software/python/simple_pendulum/trajectory_optimization/ilqr/ilqr_sympy.py
@@ -1,18 +1,18 @@
 # large parts taken from https://github.com/RussTedrake/underactuated
 
 import numpy as np
-import pydrake.symbolic as sym
+import sympy as smp
 
 
 class iLQR_Calculator():
     '''
     Class to calculate an optimal trajectory with an iterative
-    linear quadratic regulator (iLQR). This implementation uses the pydrake symbolic library.
+    linear quadratic regulator (iLQR). This implementation uses sympy.
     '''
     def __init__(self, n_x=2, n_u=1):
-        '''
+        """
         Class to calculate an optimal trajectory with an iterative
-        linear quadratic regulator (iLQR). This implementation uses the pydrake symbolic library.
+        linear quadratic regulator (iLQR). This implementation uses sympy.
 
         Parameters
         ----------
@@ -20,20 +20,23 @@ class iLQR_Calculator():
             The size of the state space.
         n_u : int, default=1
             The size of the control space.
-        '''
+        """
         self.n_x = n_x
         self.n_u = n_u
 
+        self.x_sym = smp.symbols("x:"+str(self.n_x))
+        self.u_sym = smp.symbols("u:"+str(self.n_u))
+
     def set_start(self, x0):
-        '''
+        """
         Set the start state for the trajectory.
 
         Parameters
         ----------
         x0 : array-like
             the start state. Should have the shape of (n_x,)
-        '''
-        self.x0 = np.asarray(x0)
+        """
+        self.x0 = x0
 
     def set_discrete_dynamics(self, dynamics_func):
         '''
@@ -46,7 +49,7 @@ class iLQR_Calculator():
         '''
         self.discrete_dynamics = dynamics_func
 
-    def _rollout(self, u_trj):
+    def rollout(self, u_trj):
         x_trj = np.zeros((u_trj.shape[0]+1, self.x0.shape[0]))
         x = self.x0
         i = 0
@@ -81,7 +84,7 @@ class iLQR_Calculator():
         '''
         self.final_cost = final_cost_func
 
-    def _cost_trj(self, x_trj, u_trj):
+    def cost_trj(self, x_trj, u_trj):
         total = 0.0
         ln = 0.0
         N = x_trj.shape[0]
@@ -95,103 +98,112 @@ class iLQR_Calculator():
         """
         Initialize the derivatives of the dynamics.
         """
-        self.x_sym = np.array([sym.Variable("x_{}".format(i))
-                               for i in range(self.n_x)])
-        self.u_sym = np.array([sym.Variable("u_{}".format(i))
-                               for i in range(self.n_u)])
         x = self.x_sym
         u = self.u_sym
 
-        l_stage = self.stage_cost(x, u)
-        self.l_x = sym.Jacobian([l_stage], x).ravel()
-        self.l_u = sym.Jacobian([l_stage], u).ravel()
-        self.l_xx = sym.Jacobian(self.l_x, x)
-        self.l_ux = sym.Jacobian(self.l_u, x)
-        self.l_uu = sym.Jacobian(self.l_u, u)
+        l = self.stage_cost(x, u)
+
+        l_x = smp.Matrix([l]).jacobian(x)
+        self.l_x = smp.lambdify([x, u], l_x, "numpy")
+
+        l_u = smp.Matrix([l]).jacobian(u)
[…]
