### 2021-10-29 #3853 @gkthiruvathukal
@ricopicone  Thanks for being willing to help! Let me figure out whether you need to be registered as a reviewer or not. Thanks for sharing your profile!

@openjournals/joss-eics   Am I able to add @ricopicone as a reviewer? Or does he need to be added to our list of people first?
