### 2021-11-01 #3853 @kthyng
@gkthiruvathukal and @ricopicone Anyone with access to github can review regardless of the reviewer list (pending COI concerns as appropriate of course). If you want to sign up, here is the means to do so: https://joss.theoj.org/reviewer-signup.html. Otherwise please proceed!
