### 2021-11-02 #3853 @ricopicone
@gkthiruvathukal will whedon generate a reviewer checklist like I see [here](https://github.com/openjournals/joss-reviews/issues/3783)? I have reviewed the review criteria and checklist from the JOSS docs and can begin reviewing.
