### 2021-11-02 #3884 @whedon
[REVIEW]: Torque-limited simple pendulum: A toolkit for getting familiar with control algorithms in underactuated robotics
**Submitting author:** @fwiebe (<a href="http://orcid.org/0000-0002-2195-9697">Felix Wiebe</a>)
**Repository:** https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
