### 2021-11-03 222c610
update of the main.py api for real system experiments, update setup.py
