+++ software/python/simple_pendulum/controllers/motor_control_loop.py
@@ -4,22 +4,54 @@ import numpy as np
 # driver for t-motors AK80-6
 from motor_driver.canmotorlib import CanMotorController
 
+"""
+The motor control loop only contains the minimum of code necessary to 
+send commands to and receive measurement data from the motor control 
+board in real time over CAN bus. It specifies the outgoing CAN port and 
+the CAN ID of the motor on the CAN bus and transfers this information to 
+the motor driver. It furthermore requires the following arguments:
+    control method = calls the controller, which executes the respective 
+                     control policy and returns the input torques
+    name =      needed for print outs
+    attribute = needed to differentiate between open and closed loop 
+                methods.
+    params =    parameter stored in .yaml files, although most 
+                parameters like gravity constant, link length and gear 
+                ratio, remain the same for all controllers some 
+                parameter are controller specific like e.g. reward type, 
+                integrator or learning rate for the Soft Actor Critic 
+                Controller
+    data_dict = the data dictionary contains position, velocity and 
+                torque data for each time step of the trajectory. It 
+                includes commanded as well as measured data.
+The return values start, end and meas_dt are required to monitor if 
+desired and measured time steps match.
+"""
 
-def ak80_6(control_method, name, attribute, params, data_dict):
-    motor_id = 0x01
-    can_port = 'can0'
+def ak80_6(control_method, name, attribute, params, data_dict,
+           motor_id="0x01", can_port='can0'):
+    if motor_id == "0x00":
+        motor_id = 0x00
+    elif motor_id == "0x01":
+        motor_id = 0x01
+    elif motor_id == "0x02":
+        motor_id = 0x02
+    elif motor_id == "0x03":
+        motor_id = 0x03
+    elif motor_id == "0x04":
+        motor_id = 0x04
+    # can_port = 'can0'
 
     # load dictionary entries
     n = data_dict['n']
     dt = data_dict['dt']
 
-    # load parameters
-    kp = params['kp']
-    kd = params['kd']
-
-    if name == "Feedforward Torque":
-        kp = 0
-        kd = 0
+    # load kp and kd parameters
+    kp = 0
+    kd = 0
+    if attribute == "pd_control":
+        kp = params['kp']
+        kd = params['kd']
     control_tau_max = params['torque_limit']
 
     # limit torque input to max. actuator torque
@@ -45,7 +77,7 @@ def ak80_6(control_method, name, attribute, params, data_dict):
     print()
     print("Executing", name)
     print()
-    print("Control type = ", attribute)
+    #print("Control type = ", attribute)
     print("Torque limit is set to: ", control_tau_max)
     print("kp = ", kp)
     print("kd = ", kd)
@@ -65,67 +97,67 @@ def ak80_6(control_method, name, attribute, params, data_dict):
         start_loop = time.time()
         meas_time += meas_dt
 
-        if attribute == "open_loop":
+#        if attribute == "open_loop":
+#            # get control input
+#            des_pos, des_vel, des_tau = control_method.get_control_output()
+#
+#            # clip max.torque for safety
+#            if des_tau > control_tau_max:
+#                des_tau = control_tau_max
+#            if des_tau < -control_tau_max:
+#                des_tau = -control_tau_max
+#
+#            # send control input to the actuator
+#            meas_pos, meas_vel, meas_tau = motor_01.send_rad_command(
+#                des_pos, des_vel, kp, kd, des_tau)
+#
+#            # record data
+#            data_dict["meas_pos_list"][i] = meas_pos
+#            data_dict["meas_vel_list"][i] = meas_vel
+#            data_dict["meas_tau_list"][i] = meas_tau
[…]
