+++ software/python/examples_real_system/main.py
@@ -9,20 +9,44 @@ import numpy as np
 from simple_pendulum.model.parameters import get_params
 from simple_pendulum.utilities import parse, plot, process_data
 from simple_pendulum.utilities.performance_profiler import profiler
+from simple_pendulum.utilities.process_data import prepare_trajectory
 from simple_pendulum.controllers import motor_control_loop
 from simple_pendulum.controllers.open_loop.open_loop import OpenLoopController
 from simple_pendulum.controllers.gravity_compensation.gravity_compensation import GravityCompController
 from simple_pendulum.controllers.energy_shaping.energy_shaping_controller import EnergyShapingAndLQRController
+from simple_pendulum.controllers.ilqr.iLQR_MPC_controller import iLQRMPCController
+
 try:
-    from simple_pendulum.controllers.ilqr.iLQR_MPC_controller import iLQRMPCController
+    from simple_pendulum.controllers.tvlqr.tvlqr import TVLQRController
 except ModuleNotFoundError:
+    # drake not installed
     pass
 
 try:
     from simple_pendulum.controllers.sac.sac_controller import SacController
 except ModuleNotFoundError:
+    # pytorch not installed
     pass
 
+try:
+    from simple_pendulum.controllers.ddpg.ddpg_controller import ddpg_controller
+except ModuleNotFoundError:
+    # tensorflow not installed
+    pass
+
+"""
+All parameters of the controllers in this script for the real simple pendulum
+are stored in a .yaml file. The yaml files for the different controllers
+can be found in the /data/parameters directory.
+Some parameters like (mass, length) can be measured directly, others 
+are obtained from system identification (damping, coulomb friction, 
+inertia) or depend on actuator properties (torque limits, gear ratio, 
+kp, kd).
+"""
+
+# set motor parameters
+motor_id = "0x02"
+can_port = 'can0'
 
 # run syntax parser
 args, unknown = parse.syntax()
@@ -30,21 +54,21 @@ args, unknown = parse.syntax()
 # set your workspace
 WORK_DIR = Path(Path(os.path.abspath(__file__)).parents[3])
 print("Workspace is set to:", WORK_DIR)
-sys.path.append(f'{WORK_DIR}/software/python')  # add parent folder to system path
 
 # get a timestamp
 TIMESTAMP = datetime.now().strftime("%Y%m%d-%I%M%S-%p")
 
 # select control method
-if args.openloop:
-    attribute = "open_loop"
+
+if args.ddp or args.dircol or args.ilqr:
     """
         Open loop control methods replay a precomputed trajectory. The 
         trajectories are derived offline from one out of two trajectory 
         optimization techniques:
-            - Direct Collocation (within the open source software pyDrake)
             - Feasibility-Driven Dynamic Programming (within the open source 
               software crocoddyl)
+            - Direct Collocation (within the open source software pyDrake)
+            - iterative linear quadratic regulator
         
         A trajectory is split into time steps and stored as csv file in the 
         form of position, velocity and torque data for every time step. It is 
@@ -59,12 +83,63 @@ if args.openloop:
         step size of the precomputed trajectory and a error print out that tells 
         us, if the control loop is slower then the desired time step size.
     """
+    if args.ddp:
+        """
+        This option uses a trajectory obtained via Feasibility-Driven 
+        Differential Dynamic Programming with crocoddyl. 
+        """
+        csv_file = "trajectory.csv"
+        csv_path = os.path.join(WORK_DIR, 'data', 'trajectories', 'ddp', csv_file)
+        name = "Differential Dynamic Programming"
+        folder_name = "ddp"
+
+    if args.dircol:
+        """
[…]
