+++ software/python/examples_real_system/performance_profiler.py
@@ -1,11 +1,11 @@
 from motor_driver.canmotorlib import CanMotorController
 
-from simple_pendulum.utilities.performance_profiler import motor_speed_test
+from simple_pendulum.utilities.performance_profiler import motor_speed_test, motor_send_n_commands
 
 
 # Motor ID
-motor_id = 0x01
+motor_id = 0x02
 can_port = 'can0'
 motor_1 = CanMotorController(can_port, motor_id)
 
-motor_speed_test(motor_1, n=1000)
+motor_speed_test(motor_id=motor_id, can_port=can_port, n=1000)
