+++ software/python/examples_real_system/can_motorlib_testing.py
@@ -4,7 +4,7 @@ import numpy as np
 from motor_driver.canmotorlib import CanMotorController
 
 # Motor ID
-motor_id = 0x01
+motor_id = 0x02
 
 if len(sys.argv) != 2:
     print('Provide CAN device name (can0, slcan0 etc.)')

[…]
