### 2021-12-06 048ff3e
fixed mistake in lqr controller B matrix
