### 2021-12-07 393f0b0
addressed the comments from @ricopicone in the paper
