### 2021-12-08 514fe6d
added information about encoder in the hardware description
