### 2022-01-17 b66aba8
switched to pip installation for drake and crocoddyl.
