### 2022-01-17 f0d24ba
added a readme file for example scripts
