+++ software/python/examples/README.md
@@ -1,0 +1,146 @@
+#  Examples (simulation)
+
+This folder contains example scripts for the implemented trajectory optimization, reinforcement learning and control methods.
+
+## Trajectory Optimization
+
+### Differential Dynamic Programming (DDP)
+
+Trajectory optimization with DDP is explained [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/trajectory_optimization/ddp).
+
+It can be tested with
+
+    python compute_BOXFDDP.py
+
+The script computes a trajectory and stores the trajectory in a csv file at "log_data/ddp/trajectory.csv". The trajectory is also simulated as open loop controller with pinocchio and visualized with the gepetto viewer.
+
+Requirements: Crocoddyl
+
+### Direct Collocation
+
+Trajectory optimization with direct collocation is explained [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/trajectory_optimization/direct_collocation). 
+
+It can be tested with
+
+    python compute dircol_swingup.py
+
+The script computes a trajectory and stores the trajectory in a csv file at "log_data/direct_collocation/trajectory.csv". The trajectory is also simulated with a [TVLQR](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/controllers/tvlqr) controller which stabilizes the trajectory.
+
+Requirements: Drake
+
+### Iterative Linear Quadratic Regulator (iLQR)
+
+Trajectory optimization with iLQR is explained [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/trajectory_optimization/ilqr).
+
+It can be tested with
+
+    python compute_iLQR_swingup.py
+
+The script computes a trajectory and stores the trajectory in a csv file at "log_data/ilqr/trajectory.csv". The trajectory is also simulated as open loop controller.
+
+Requirements: Drake (optional)
+
+If Drake is installed the symbolic libary of Drake will be used, else sympy is used.
+
+## Reinforcement Learning
+
+### State Actor Critic (SAC)
+
+The SAC reinforcement learning method is explained [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/controllers/sac).
+
+To train a model:
+
+    python train_sac.py
+
+This might take a while. The model will be saved at "log_data/sac_training/sac_model.zip". A trained model is provided [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/data/models/sac_model.zip). It can be tested with
+
+    python sim_sac.py
+
+This will simulate and visualize the behavior of the pendulum following this policy for 10s.
+
+Requirements: torch, stable-baselines3
+
+Note: The provided model was trained with versions: torch==1.10.0 and stable-baselines==1.3.0. Loading it with other versions may cause errors.
+
+### Deep Deterministic Policy Gradient (DDPG)
+
+The DDPG reinforcement learning method is explained [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/controllers/ddpg).
+
+To train a model:
+
+    python train_ddpg.py
+
+This might take a while. The model will be saved at "log_data/ddpg_training/ddpg_model". A trained model is provided [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/data/models/ddpg_model). It can be tested with
+
+    python sim_ddpg.py
+
+This will simulate and visualize the behavior of the pendulum following this policy for 10s.
+
[…]
