+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 
 @article{hargraves1987direct,
-  title={Direct trajectory optimization using nonlinear programming and collocation},
+  title={{Direct Trajectory Optimization Using Nonlinear Programming and Collocation}},
   author={Hargraves, Charles R and Paris, Stephen W},
   journal={Journal of guidance, control, and dynamics},
   volume={10},
@@ -17,7 +17,8 @@
 }
 
 @misc{haarnoja2018soft,
-      title={Soft Actor-Critic: Off-Policy Maximum Entropy Deep Reinforcement Learning with a Stochastic Actor}, 
+      title={{Soft Actor-Critic: Off-Policy Maximum Entropy Deep Reinforcement Learning with a 
+Stochastic Actor}}, 
       author={Tuomas Haarnoja and Aurick Zhou and Pieter Abbeel and Sergey Levine},
       year={2018},
       eprint={1801.01290},
@@ -26,7 +27,7 @@
 }
 
 @misc{lillicrap2019continuous,
-      title={Continuous control with deep reinforcement learning}, 
+      title={{Continuous Control with Deep Reinforcement Learning}}, 
       author={Timothy P. Lillicrap and Jonathan J. Hunt and Alexander Pritzel and Nicolas Heess and Tom Erez and Yuval Tassa and David Silver and Daan Wierstra},
       year={2019},
       eprint={1509.02971},
@@ -37,7 +38,7 @@
 @INPROCEEDINGS{mastalli2020crocoddyl,
   author={Mastalli, Carlos and Budhiraja, Rohan and Merkt, Wolfgang and Saurel, Guilhem and Hammoud, Bilal and Naveau, Maximilien and Carpentier, Justin and Righetti, Ludovic and Vijayakumar, Sethu and Mansard, Nicolas},
   booktitle={2020 IEEE International Conference on Robotics and Automation (ICRA)}, 
-  title={Crocoddyl: An Efficient and Versatile Framework for Multi-Contact Optimal Control}, 
+  title={{Crocoddyl: An Efficient and Versatile Framework for Multi-Contact Optimal Control}}, 
   year={2020},
   volume={},
   number={},
@@ -45,16 +46,16 @@
   doi={10.1109/ICRA40945.2020.9196673}}
 
 @misc{drake,
- author = "Russ Tedrake and the Drake Development Team",
- title = "Drake: Model-based design and verification for robotics",
- year = 2019,
- url = "https://drake.mit.edu"
+ author = {Russ Tedrake and the Drake Development Team},
+ title = {{Drake: Model-Based Design and Verification for Robotics}},
+ year = {2019},
+ url = {https://drake.mit.edu}
 }
 
 @misc{stable-baselines3,
   author = {Raffin, Antonin and Hill, Ashley and Ernestus, Maximilian and Gleave, Adam and 
 Kanervisto, Anssi and Dormann, Noah},
-  title = {Stable Baselines3},
+  title = {{Stable Baselines3}},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -62,7 +63,7 @@ Kanervisto, Anssi and Dormann, Noah},
 }
 
 @misc{tensorflow2015-whitepaper,
-title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+title={ {TensorFlow}: {Large-Scale Machine Learning on Heterogeneous Systems}},
 url={https://www.tensorflow.org/},
 note={Software available from tensorflow.org},
 author={
@@ -111,7 +112,8 @@ author={
 
 
 @inproceedings{bargsten2016,
[…]
