### 2022-03-06 6caf96f
changed control categories in figure, readme and paper
