+++ software/simple_pendulum/analysis/benchmark.py
@@ -1,0 +1,486 @@
+import time
+import yaml
+import numpy as np
+
+from simple_pendulum.model.pendulum_plant import PendulumPlant
+from simple_pendulum.simulation.simulation import Simulator
+
+
+"""
+Benchmarks
+==========
+"""
+
+
+def modify_pendulum_parameter(par):
+    mod_par = np.copy(par)
+    mod_par += np.random.rand()*0.1
+    mod_par *= np.random.rand()*1.0 + 0.5
+    return mod_par
+
+
+class benchmarker():
+    def __init__(self,
+                 dt=0.01,
+                 max_time=10.0,
+                 integrator="runge_kutta",
+                 benchmark_iterations=10):
+
+        self.dt = dt
+        self.max_time = max_time
+        self.max_steps = int(self.max_time / self.dt)
+        self.integrator = "runge_kutta"
+        self.iterations = benchmark_iterations
+
+        self.x0 = np.array([0.0, 0.0])
+        self.goal = np.array([np.pi, 0.0])
+        self.epsilon = np.array([0.1, 0.1])
+
+    def init_pendulum(self, mass=0.57288, length=0.5, inertia=None,
+                      damping=0.15, coulomb_friction=0.0, gravity=9.81,
+                      torque_limit=2.0):
+        """
+        Initialize the pendulum parameters.
+
+        Parameters
+        ----------
+        mass : float, default=0.57288
+            mass of the pendulum [kg]
+        length : float, default=0.5
+            length of the pendulum [m]
+        inertia : float, default=None
+            inertia of the pendulum [kg m^2]
+            defaults to point mass inertia (mass*length^2)
+        damping : float, default=0.15
+            damping factor of the pendulum [kg m/s]
+        coulomb_friction : float, default=0.0
+            coulomb friciton of the pendulum [Nm]
+        gravity : float, default=9.81
+            gravity (positive direction points down) [m/s^2]
+        torque_limit : float, default=2.0
+            the torque_limit of the pendulum actuator
+        """
+
+        self.pen_mass = mass
+        self.pen_length = length
+        if inertia is None:
+            inertia = mass*length**2
+        self.pen_inertia = inertia
+        self.pen_damping = damping
+        self.pen_cfric = coulomb_friction
+        self.pen_gravity = gravity
+        self.pen_torque_limit = torque_limit
+
+        self.pendulum = PendulumPlant(mass=self.pen_mass,
+                                      length=self.pen_length,
+                                      damping=self.pen_damping,
+                                      gravity=self.pen_gravity,
+                                      coulomb_fric=self.pen_cfric,
+                                      inertia=self.pen_inertia,
+                                      torque_limit=self.pen_torque_limit)
+
+        self.ref_pendulum = PendulumPlant(mass=self.pen_mass,
+                                          length=self.pen_length,
+                                          damping=self.pen_damping,
+                                          gravity=self.pen_gravity,
+                                          coulomb_fric=self.pen_cfric,
+                                          inertia=self.pen_inertia,
+                                          torque_limit=self.pen_torque_limit)
+
+        self.simulator = Simulator(plant=self.pendulum)
+        self.ref_simulator = Simulator(plant=self.ref_pendulum)
+
+    def set_controller(self, controller):
+        self.controller = controller
+
+    def check_regular_execution(self):
+
+        t0 = 0.0
+        self.controller.set_goal(x=self.goal)
+        self.controller.init(x0=self.x0)
+
+        x = np.copy(self.x0)
+        t = np.copy(t0)
+        self.simulator.reset_data_recorder()
+        self.ref_simulator.reset_data_recorder()
+        self.simulator.set_state(time=t, x=np.copy(x))
+
+        tau_list = []
[…]
