+++ software/simple_pendulum/controllers/tvlqr/__init__.py
@@ -1,4 +1,4 @@
 """
-Time-varying Linear Quadrativ Regulator (TVLQR)
+Time-varying Linear Quadratic Regulator (TVLQR)
 ===============================================
 """
