+++ software/python/simple_pendulum/controllers/lqr/roa/utils.py
@@ -1,0 +1,91 @@
+from scipy.spatial.transform import Rotation as R
+from scipy import linalg
+from scipy.special import gamma, factorial
+import numpy as np
+
+
+
+def direct_sphere(d,r_i=0,r_o=1):
+    """Direct Sampling from the d Ball based on Krauth, Werner. Statistical Mechanics: Algorithms and Computations. Oxford Master Series in Physics 13. Oxford: Oxford University Press, 2006. page 42
+
+    Parameters
+    ----------
+    d : int
+        dimension of the ball
+    r_i : int, optional
+        inner radius, by default 0
+    r_o : int, optional
+        outer radius, by default 1
+
+    Returns
+    -------
+    np.array
+        random vector directly sampled from the solid d Ball
+    """
+    # vector of univariate gaussians:
+    rand=np.random.normal(size=d)
+    # get its euclidean distance:
+    dist=np.linalg.norm(rand,ord=2)
+    # divide by norm
+    normed=rand/dist
+    
+    # sample the radius uniformly from 0 to 1 
+    rad=np.random.uniform(r_i,r_o**d)**(1/d)
+    # the r**d part was not there in the original implementation.
+    # I added it in order to be able to change the radius of the sphere
+    # multiply with vect and return
+    return normed*rad
+
+def sample_from_ellipsoid(M,rho,r_i=0,r_o=1):
+    """sample directly from the ellipsoid defined by xT M x.
+
+    Parameters
+    ----------
+    M : np.array
+        Matrix M such that xT M x leq rho defines the hyperellipsoid to sample from
+    rho : float
+        rho such that xT M x leq rho defines the hyperellipsoid to sample from
+    r_i : int, optional
+        inner radius, by default 0
+    r_o : int, optional
+        outer radius, by default 1
+
+    Returns
+    -------
+    np.array
+        random vector from within the hyperellipsoid
+    """
+    lamb,eigV=np.linalg.eigh(M/rho) 
+    d=len(M)
+    xy=direct_sphere(d,r_i=r_i,r_o=r_o) #sample from outer shells
+    T=np.linalg.inv(np.dot(np.diag(np.sqrt(lamb)),eigV.T)) #transform sphere to ellipsoid (refer to e.g. boyd lectures on linear algebra)
+    return np.dot(T,xy.T).T
+
+def quad_form(M,x):
+    """
+    Helper function to compute quadratic forms such as x^TMx
+    """
+    return np.dot(x,np.dot(M,x))
+
+
+def vol_ellipsoid(rho,M):
+    """
+    Calculate the Volume of a Hyperellipsoid
+    Volume of the Hyperllipsoid according to https://math.stackexchange.com/questions/332391/volume-of-hyperellipsoid/332434
+    Intuition: https://textbooks.math.gatech.edu/ila/determinants-volumes.html
+    Volume of n-Ball https://en.wikipedia.org/wiki/Volume_of_an_n-ball
+    """
+    
+    # For a given hyperellipsoid, find the transformation that when applied to the n Ball yields the hyperellipsoid
+    lamb,eigV=np.linalg.eigh(M/rho) 
+    A=np.dot(np.diag(np.sqrt(lamb)),eigV.T) #transform ellipsoid to sphere
+    detA=np.linalg.det(A)
+    
+    # Volume of n Ball (d dimensions)
+    d=M.shape[0] # dimension 
+    volC=(np.pi**(d/2))/(gamma((d/2)+1))
+
+    # Volume of Ellipse
+    volE=volC/detA
+
+    return volE
