### 2022-03-29 88415bf
Added SOS methodology to RoA estimation
