+++ software/python/simple_pendulum/controllers/lqr/roa/utils.py
@@ -2,6 +2,7 @@ from scipy.spatial.transform import Rotation as R
 from scipy import linalg
 from scipy.special import gamma, factorial
 import numpy as np
+from pydrake.all import (MathematicalProgram, Solve, Variables, Jacobian)
 
 
 
@@ -89,3 +90,74 @@ def vol_ellipsoid(rho,M):
     volE=volC/detA
 
     return volE
+
+def rhoVerification(rho, pendulum, controller):
+    """SOS Verification of the Lyapunov conditions for a given rho in order to obtain an estimate  of the RoA for the closed loop dynamics.
+       This method is described by Russ Tedrake in "Underactuated Robotics: Algorithms for Walking, Running, Swimming, Flying, and Manipulation", 
+       Course Notes for MIT 6.832, 2022, "http://underactuated.mit.edu", sec. 9.3.2: "Basic region of attraction formulation".
+
+    Parameters
+    ----------
+    rho: float
+        value of rho to be verified
+    pendulum : simple_pendulum.model.pendulum_plant
+        configured pendulum plant object
+    controller : simple_pendulum.controllers.lqr.lqr_controller
+        configured lqr controller object
+
+    Returns
+    -------
+    result : boolean
+        result of the verification
+    """
+
+    #K and S matrices from LQR control
+    K = controller.K
+    S = controller.S
+
+    # Pendulum parameters
+    m = pendulum.m
+    l = pendulum.l
+    g = pendulum.g
+    b = pendulum.b
+    torque_limit = pendulum.torque_limit
+
+    # non-linear dyamics
+    prog = MathematicalProgram()
+    xbar = prog.NewIndeterminates(2, "x")
+    xg = [np.pi, 0]  # reference
+    x = xbar + xg
+    ubar = -K.dot(xbar)[0] # control input with reference
+    Tsin = -(x[0]-xg[0]) + (x[0]-xg[0])**3/6 - (x[0]-xg[0])**5/120 + (x[0]-xg[0])**7/5040
+    fn = [x[1], (ubar-b*x[1]-Tsin*m*g*l)/(m*l*l)]
+
+    # cost-to-go of LQR as Lyapunov candidate
+    V = (xbar).dot(S.dot(xbar))
+    Vdot = Jacobian([V], xbar).dot(fn)[0]
+
+    # Saturation for fn and Vdot
+    u_minus = - torque_limit
+    u_plus = torque_limit
+    fn_minus = [x[1], (u_minus-b*x[1]-Tsin*m*g*l)/(m*l*l)] 
+    Vdot_minus = V.Jacobian(xbar).dot(fn_minus)
+    fn_plus = [x[1], (u_plus-b*x[1]-Tsin*m*g*l)/(m*l*l)] 
+    Vdot_plus = V.Jacobian(xbar).dot(fn_plus)
+
+    # Define the Lagrange multipliers.
+    lambda_1 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
+    lambda_2 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
+    lambda_3 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
+    lambda_4 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
+    lambda_5 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
+    lambda_6 = prog.NewSosPolynomial(Variables(xbar), 6)[0].ToExpression()
[…]
