### 2022-03-29 ac05f47
Added Verification script for RoA estimation
