### 2022-03-29 c8bc585
added author (CubeMars) for tmotors manual
