+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{weiwei2004iterative,
-  	Author = {{Weiwei}, L.. and {Todorov}, E.},
+  	author = {{Weiwei}, L.. and {Todorov}, E.},
   	Journal = {International Conference on Informatics in Control, Automation and Robotics},
   	Title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
   	Year = 2004,
@@ -174,7 +174,7 @@ Techniques}},
 }
 
 @misc{tmotors_manual,
-  author = {},
+  author = {CubeMars},
   year = {2021},
   title = {{AK Series Actuator Driver Manual}},
   url = {{https://store.cubemars.com/images/file/20211201/1638329381542610.pdf}},
