### 2022-03-31 4b5e422
Added analysis for the taylor-approximated dynamics
