+++ software/simple_pendulum/trajectory_optimization/ilqr/ilqr_sympy.py
@@ -1,13 +1,12 @@
-# large parts taken from https://github.com/RussTedrake/underactuated
+"""
+Symbolic
+========
+
+Large parts taken from `Russ Tedrake <https://github.com/RussTedrake/underactuated>`_.
+"""
 
 import numpy as np
 import sympy as smp
-
-
-"""
-SymPy
-=====
-"""
 
 
 class iLQR_Calculator():
