+++ software/simple_pendulum/utilities/filters/butterworth.py
@@ -1,3 +1,8 @@
+"""
+Butterworth Low-pass Filter
+===========================
+"""
+
 # import numpy as np
 import matplotlib.pyplot as plt
 from scipy import signal, optimize
