+++ software/simple_pendulum/controllers/abstract_controller.py
@@ -1,12 +1,12 @@
-# here, the abstract controller class is defined, to which all controller
-# classes have to adhere
-from abc import ABC, abstractmethod
-
-
 """
 Abstract Controller
 ===================
+
+Abstract controller class to which all controller classes have to adhere.
 """
+
+
+from abc import ABC, abstractmethod
 
 
 class AbstractController(ABC):
@@ -21,25 +21,25 @@ class AbstractController(ABC):
         Supposed to be overwritten by actual controllers. The API of this
         method should be adapted. Unused inputs/outputs can be set to None.
 
-        Parameters
-        ----------
-        meas_pos : float
-            the position of the pendulum [rad]
-        meas_vel : float
-            the velocity of the pendulum [rad/s]
-        meas_tau : float
-            the meastured torque of the pendulum [Nm]
-        meas_time : float
-            the collapsed time [s]
+        **Parameters**
+
+        ``meas_pos``: ``float``
+            The position of the pendulum [rad]
+        ``meas_vel``: ``float``
+            The velocity of the pendulum [rad/s]
+        ``meas_tau``: ``float``
+            The meastured torque of the pendulum [Nm]
+        ``meas_time``: ``float``
+            The collapsed time [s]
 
         Returns
         -------
-        des_pos : float
-            the desired position of the pendulum [rad]
-        des_vel : float
-            the desired velocity of the pendulum [rad/s]
-        des_tau : float
-            the torque supposed to be applied by the actuator [Nm]
+        ``des_pos``: ``float``
+            The desired position of the pendulum [rad]
+        ``des_vel``: ``float``
+            The desired velocity of the pendulum [rad/s]
+        ``des_tau``: ``float``
+            The torque supposed to be applied by the actuator [Nm]
         """
 
         des_pos = None
@@ -53,8 +53,8 @@ class AbstractController(ABC):
 
         Parameters
         ----------
-        x0 : array like
-            the start state of the pendulum
+        ``x0``: ``array like``
+            The start state of the pendulum
         """
         self.x0 = x0
 
@@ -64,8 +64,8 @@ class AbstractController(ABC):
 
         Parameters
         ----------
-        x : array like
-            the desired goal state of the controller
+        ``x``: ``array like``
+            The desired goal state of the controller
         """
 
         self.goal = x
