+++ software/simple_pendulum/reinforcement_learning/ddpg/replay_buffer.py
@@ -1,11 +1,11 @@
-import numpy as np
-import tensorflow as tf
-
-
 """
 Replay Buffer
 =============
 """
+
+
+import numpy as np
+import tensorflow as tf
 
 
 class ReplayBuffer:
