### 2022-04-01 e39a5a2
Improved plots for RoA analysis and fixed taylorApprox order problem
