+++ software/python/examples/verify_roa_estimation.py
@@ -5,7 +5,7 @@ from simple_pendulum.model.pendulum_plant import PendulumPlant
 from simple_pendulum.controllers.lqr.lqr_controller import LQRController
 from simple_pendulum.simulation.simulation import Simulator
 
-from simple_pendulum.controllers.lqr.roa.plot import get_ellipse_patch 
+from simple_pendulum.controllers.lqr.roa.plot import get_ellipse_patch
 
 from simple_pendulum.controllers.lqr.roa.sampling import najafi_based_sampling
 from simple_pendulum.controllers.lqr.roa.sos import SOSequalityConstrained, SOSlineSearch
@@ -14,8 +14,8 @@ from simple_pendulum.controllers.lqr.roa.utils import sample_from_ellipsoid
 roa_estimation_methods  = [najafi_based_sampling, SOSequalityConstrained, SOSlineSearch]
 linestyles              = ["-",":","--"]
 
-torque_limit = 8
-init_num = 1000
+torque_limit = 4 # tactuator orque limit
+init_num = 500 # number of random initial conditions to check
 
 mass = 0.57288
 length = 0.5
@@ -25,12 +25,12 @@ coulomb_fric = 0.0
 inertia = mass*length*length
 
 
-fig, ax = plt.subplots()
-plt.scatter([np.pi],[0],color="black",marker="x")
-plt.annotate("Goal state",(np.pi,0))
+fig, ax = plt.subplots(figsize=(18, 8))
+x_g = plt.scatter([np.pi],[0],color="black",marker="x", linewidths=3)
 
 rho_max = 0
-
+p_buffer = []
+# RoA estimation with the three different methods
 for j,mthd in enumerate(roa_estimation_methods):
 
     pendulum = PendulumPlant(mass=mass,
@@ -57,7 +57,9 @@ for j,mthd in enumerate(roa_estimation_methods):
     if (rho != 0):
         p = get_ellipse_patch(np.pi,0,rho,S,linec= "black",linest=linestyles[j])
         ax.add_patch(p)
+        p_buffer = np.append(p_buffer,p)
 
+# Sampling and check of initial conditions inside the biggest founded ellipse
 for i in range(0,init_num):   
     x_i = sample_from_ellipsoid(S,rho_max)
 
@@ -73,20 +75,24 @@ for i in range(0,init_num):
                     controller=controller,
                     integrator="runge_kutta"
                     )
-        
-    if (round(np.asarray(X).T[0][-1]) == round(np.pi) and round(np.asarray(X).T[1][-1]) == 0):
-        plt.scatter([x_i[0]+ np.pi],[x_i[1]],color="green",marker="o")
-    else:
-        plt.scatter([x_i[0]+ np.pi],[x_i[1]],color="red",marker="o")
 
+    # coloring the checked initial states depending on the result    
+    if (round(np.asarray(X).T[0][-1]) == round(np.pi) and round(np.asarray(X).T[1][-1]) == 0):
+        greenDot = plt.scatter([x_i[0]+ np.pi],[x_i[1]],color="green",marker="o")
+        redDot = None
+    else:
+        redDot = plt.scatter([x_i[0]+ np.pi],[x_i[1]],color="red",marker="o")
+
+ax.set_xlabel("x")
+ax.set_ylabel(r"$\dot{x}$")
+if (not redDot == None):
+    ax.legend(handles = [greenDot,redDot,x_g,p_buffer[0],p_buffer[1],p_buffer[2]], 
+                labels = ["successfull initial state","failing initial state", "Goal state", 
+                "najafi-based sampling method", "SOS method wirh equality-constrained formulation", "SOS method with line search"])
+else: 
[…]
