+++ software/python/examples/taylorApprox_roa_sos.py
@@ -23,6 +23,7 @@ controller = LQRController(mass=mass,
                            gravity=gravity,
                            torque_limit=torque_limit)
 
+# non-linear dynamics simulation
 pendulum = PendulumPlant(mass=mass,
                          length=length,
                          damping=damping,
@@ -47,24 +48,41 @@ theta = np.asarray(X).T[0]
 theta_dot = np.asarray(X).T[1]
 theta_Ddot = (U- mass* gravity * length * np.sin(theta) - damping * theta_dot) / inertia
 
+# choice of the max order for the Taylor expansion
+max_order = 7
+orders = range(1,max_order+1,2)
+
 fig, ax = plt.subplots(3, 1, figsize=(18, 8), sharex="all")
 ax[0].plot(T, theta, label="non-linear", linewidth = 5, color = "black", alpha = 0.2)
 ax[0].set_ylabel("angle [rad]")
 ax[1].plot(T, theta_dot, label="non-linear", linewidth = 5,color = "black", alpha = 0.2)
 ax[1].set_ylabel("angular velocity [rad/s]")
 ax[2].plot(T, theta_Ddot, label="non-linear", linewidth = 5,color = "black", alpha = 0.2)
-ax[2].set_ylabel("angular accelleration [rad/s^2]")
+ax[2].set_ylabel(r"angular accelleration [rad/$s^2$]")
 ax[2].set_xlabel("time [s]")
+plt.suptitle("Simulated Taylor approximations of the closed-loop dynamics")
 
-max_order = 4
-err = []
-for j in range(1,max_order+1):                                
+# dynamic figure structure
+o_c_max = 0
+if (len(orders)%2 ==0):
+    o_c_max = round(len(orders)/2)
+    fig1, ax1 = plt.subplots(2, o_c_max, figsize=(len(orders)*3, len(orders)*3), sharex="all")
+else: 
+    o_c_max = round(len(orders)/2 + len(orders)%2)
+    fig1, ax1 = plt.subplots(2, o_c_max, figsize=(len(orders)*3, len(orders)*3), sharex="all")
+
+plt.suptitle("Taylor approximations of the closed-loop dynamics: State dependent error")
+
+o_r = 0
+o_c = 0
+for k in orders:  
+    # k-order Taylor approximated dynamics simulation                           
     pendulum_approx = PendulumPlantApprox(mass=mass,
                                       length=length,
                                       damping=damping,
                                       gravity=gravity,
                                       torque_limit=torque_limit,
-                                      taylorApprox_order = j)
+                                      taylorApprox_order = k)
 
     sim_approx = Simulator(plant=pendulum_approx)
     T_approx, X_approx, U_approx = sim_approx.simulate(t0=0.0,
@@ -73,25 +91,48 @@ for j in range(1,max_order+1):
                                                     dt=dt,
                                                     controller=controller,
                                                     integrator="runge_kutta"
-                                                    )
-
-    
+                                                    )    
     
     theta_approx = np.asarray(X_approx).T[0]
     theta_dot_approx = np.asarray(X_approx).T[1]
 
     x0 = Variable("theta")
-    Tsin_exp = TaylorExpand(np.sin(x0), {x0: np.pi},j)
+    Tsin_exp = TaylorExpand(np.sin(x0), {x0: np.pi}, k)
     Tsin = np.empty(len(theta_approx))
[…]
