+++ .github/workflows/pytest.yml
@@ -89,6 +89,7 @@ jobs:
     # Run unit tests =============================================
     - name: Run tests
       if: env.run_tests == 'true'
+      id: tests
       run: |
         # Set environment variables to ensure dependencies are found
         export PATH=/opt/openrobots/bin:$PATH
@@ -98,28 +99,32 @@ jobs:
         # Run tests
         cd $codedir
         coverage run -m pytest
+      continue-on-error: true
     - name: Coverage
-      run: |      
+      run: |
         cd $codedir
         # Report
         coverage report
         # Emblem
-        emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
-        ls
-        cd ..
-        ls        
+        if ${{ steps.tests.outcome == 'success' }}
+        then
+          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --colors "#940000" "#6bd8ff" --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
+        else
+          emblem Error --label "Py3.6 Ub18.04" --colors "#940000" "#6bd8ff" --fname ./ub20cov.svg
+        fi
     - name: Set
       if: env.run_tests == 'true'
       run: |
-        mv -v $codedir/ub18cov.svg ./cov/ub18cov.svg
+        mkdir --parents cov
+        mv -v $codedir/ub18cov.svg $_
     - name: Push
       if: env.branch != ''
       run: |
         git config --local user.email "41898282+github-actions[bot]@users.noreply.github.com"
         git config --local user.name "github-actions[bot]"
         git pull
-        git add ./cov/ub18cov.svg && git commit --allow-empty -m "Updated Documentation"
-        git push -u origin $branch
+        git add cov/ub18cov.svg && git commit --allow-empty -m "Badge updated"
+        git push -u
   # ===========================
   #
   #         Ubuntu 20
@@ -156,8 +161,6 @@ jobs:
       run: |
         pip install coverage
         pip install emblem
-        emblem 80%
-        ls -a
     - name: Install Python dev headers
       if: env.run_tests == 'true'
       run: |
@@ -196,6 +199,7 @@ jobs:
     # Run unit tests =============================================
     - name: Run tests
       if: env.run_tests == 'true'
+      id: tests
       run: |
         # Set environment variables to ensure dependencies are found
         export PATH=/opt/openrobots/bin:$PATH
@@ -205,22 +209,29 @@ jobs:
         # Run tests
         cd $codedir
         coverage run -m pytest
+      continue-on-error: true
     - name: Coverage
-      run: |      
+      run: |
         cd $codedir
         # Report
         coverage report
         # Emblem
-        emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+        if ${{ steps.tests.outcome == 'success' }}
+        then
+          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --colors "#940000" "#6bd8ff" --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+        else
+          emblem --label "Py3.8 Ub20.04" --colors "#940000" "#6bd8ff" --fname ./ub20cov.svg
+        fi
     - name: Set
       if: env.run_tests == 'true'
       run: |
-        mv -v $codedir/ub20cov.svg ./cov/ub20cov.svg
+        mkdir --parents cov
+        mv -v $codedir/ub20cov.svg $_
     - name: Push
       if: env.branch != ''
       run: |
         git config --local user.email "41898282+github-actions[bot]@users.noreply.github.com"
[…]
