+++ README.md
@@ -8,12 +8,11 @@
 <img width="300" src="docs/reference/source/figures/pendulum_light_painting.jpg">
 </div>
 
-<div align="center">
-<img width="300" src="cov/ub18cov.svg">
-<img width="300" src="cov/ub20cov.svg">
-</div>
-
 ## Introduction #
+<div align="left">
+<img src="cov/ub18cov.svg">
+<img src="cov/ub20cov.svg">
+</div>
 The project is an open-source and low-cost kit to get started with underactuated robotics. The kit targets lowering the entry barrier for studying underactuation in real systems which is often overlooked in conventional robotics courses. It implements a **torque-limited simple pendulum** built using a quasi-direct drive motor which allows for a low friction, torque limited setup. This project describes the _offline_ and _online_ control methods which can be studied using the kit, lists its components, discusses best practices for implementation, presents results from experiments with the simulator and the real system. This repository describes the hardware (CAD, Bill Of Materials (BOM) etc.) required to build the physical system and provides the software (URDF models, simulation and controller) to control it.
 
 **See the simple pendulum in action:** 
