+++ .github/workflows/pytest.yml
@@ -108,9 +108,9 @@ jobs:
         # Emblem
         if ${{ steps.tests.outcome == 'success' }}
         then
-          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --colors "#940000" "#6bd8ff" --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
+          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
         else
-          emblem Error --label "Py3.6 Ub18.04" --colors "#940000" "#6bd8ff" --fname ./ub20cov.svg
+          emblem Error --label "Py3.6 Ub18.04" --fname ./ub20cov.svg
         fi
     - name: Set
       if: env.run_tests == 'true'
@@ -218,9 +218,9 @@ jobs:
         # Emblem
         if ${{ steps.tests.outcome == 'success' }}
         then
-          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --colors "#940000" "#6bd8ff" --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
         else
-          emblem --label "Py3.8 Ub20.04" --colors "#940000" "#6bd8ff" --fname ./ub20cov.svg
+          emblem Error --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
         fi
     - name: Set
       if: env.run_tests == 'true'
