### 2022-04-02 f3872b3
2 fixes for automated testing workflow
