+++ .github/workflows/pytest.yml
@@ -109,7 +109,12 @@ jobs:
         # Emblem
         if ${{ steps.tests.outcome == 'success' }}
         then
-          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
+          if [$badge_content == 'cov']
+          then
+            emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
+          else
+            emblem $badge_content --color 3ade65 --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
+          fi
         else
           emblem Error --label "Py3.6 Ub18.04" --fname ./ub18cov.svg
         fi
@@ -126,6 +131,10 @@ jobs:
         git pull
         git add cov/ub18cov.svg && git commit --allow-empty -m "Badge updated"
         git push -u
+    - name: Close
+      if: steps.tests.outcome != 'success'
+      run: |
+        exit 1
   # ===========================
   #
   #         Ubuntu 20
@@ -220,7 +229,12 @@ jobs:
         # Emblem
         if ${{ steps.tests.outcome == 'success' }}
         then
-          emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+          if [$badge_content == 'cov']
+          then
+            emblem $(coverage report | awk '$1 == "TOTAL" {print $NF+0}') --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+          else
+            emblem $badge_content --color 3ade65 --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
+          fi
         else
           emblem Error --label "Py3.8 Ub20.04" --fname ./ub20cov.svg
         fi
@@ -237,3 +251,7 @@ jobs:
         git pull
         git add cov/ub20cov.svg && git commit --allow-empty -m "Badge updated"
         git push -u
+    - name: Close
+      if: steps.tests.outcome != 'success'
+      run: |
+        exit 1
