+++ docs/paper/paper.md
@@ -1,245 +1,0 @@
----
-title: 'Torque-limited simple pendulum: A toolkit for getting familiar with control algorithms in underactuated robotics'
-tags:
-  - Python
-  - robotics
-  - underactuated
-  - pendulum
-  - torque-limited
-authors:
-  - name: Felix Wiebe^[co-first author] # note this makes a footnote saying 'co-first author'
-    orcid: 0000-0002-2195-9697
-    affiliation: 1 # (Multiple affiliations must be quoted)
-  - name: Jonathan Babel^[co-first author] # note this makes a footnote saying 'co-first author'
-    affiliation: 1
-  - name: Shivesh Kumar
-    orcid: 0000-0002-6254-3882
-    affiliation: 1    
-  - name: Shubham Vyas
-    affiliation: 1
-  - name: Daniel Harnack
-    affiliation: 1
-  - name: Melya Boukheddimi
-    affiliation: 2
-  - name: Mihaela Popescu
-    affiliation: 2    
-  - name: Frank Kirchner
-    affiliation: 1, 2
-affiliations:
- - name: DFKI GmbH Robotics Innovation Center, Bremen, Germany
-   index: 1
- - name: Working Group Robotics, University of Bremen, Bremen, Germany
-   index: 2
-
-date: 04 October 2021
-bibliography: paper.bib
----
-
-# Summary
-
-There are many, wildly different approaches to robotic control.
-Underactuated robots are systems for which it is not possible to dictate arbitrary accelerations to all joints. Hence, a controller cannot be used to override the system dynamics and force the system on a desired trajectory as it often is done in classical control techniques.
-A torque-limited pendulum is arguably the simplest underactuated robotic system and thus is a suitable system to study, test and benchmark different controllers.
-
-This repository describes the hardware (Computer-aided design (CAD) models, Bill Of Materials (BOM), etc.) required to build a
-physical pendulum system and provides the software (Unified Robot Description Format (URDF) models, simulation and controller) to control it. It provides a setup for studying established and novel control methods, and targets students and researchers interested in studying underactuation.
-
-
-# Statement of need
-
-This repository is designed to be used in education and research. It lowers the entry barrier
-for studying underactuation in real systems which is often overlooked in conventional robotics courses.
-With this software package, students who want to learn about robotics, optimal control or reinforcement learning can make hands-on
-experiences with hardware and software for robot control.
-The dualistic approach of describing software and hardware as well as the large spectrum of control methods are outstanding features of this package in comparison to similar software such as stable baselines [@stable-baselines3], open AI gym [@Brockman2016] and Drake [@drake].
-Results from real experiments are provided to ensure reproducibility and evaluate novel control methods.
-
-
-# Background
-
-This project provides an easy accessible plant for the pendulum dynamics which is built up from scratch and uses only standard libraries. The plant can be passed to a simulator object, which is capable of integrating the equations of motion and thus simulating the pendulum's motion forward in time. The simulator can perform Euler and Runge-Kutta integration and can also visualize the motion in a matplotlib animation. Furthermore, it is possible to interface a controller to the simulator which sends control a signal in form of a torque $\tau$ to the motor.
-
-The pendulum has a stable (downward configuration) and unstable (upward configuration) fixed points. A challenge from the control point of view is to swing the pendulum up to the unstable fixed point and stabilize the pendulum in that state respecting the torque limitation.
-
-## Mechanical Setup
-
[…]
