+++ docs/paper/paper.bib
@@ -1,184 +1,0 @@
-@article{weiwei2004iterative,
-  	author = {{Weiwei}, L.. and {Todorov}, E.},
-  	Journal = {International Conference on Informatics in Control, Automation and Robotics},
-  	Title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
-  	Year = 2004,
-    pages = 222-229
-}
-
-@article{hargraves1987direct,
-  title={{Direct Trajectory Optimization Using Nonlinear Programming and Collocation}},
-  author={Hargraves, Charles R and Paris, Stephen W},
-  journal={Journal of guidance, control, and dynamics},
-  volume={10},
-  number={4},
-  pages={338--342},
-  year={1987}
-}
-
-@misc{haarnoja2018soft,
-      title={{Soft Actor-Critic: Off-Policy Maximum Entropy Deep Reinforcement Learning with a 
-Stochastic Actor}}, 
-      author={Tuomas Haarnoja and Aurick Zhou and Pieter Abbeel and Sergey Levine},
-      year={2018},
-      eprint={1801.01290},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
-}
-
-@misc{lillicrap2019continuous,
-      title={{Continuous Control with Deep Reinforcement Learning}}, 
-      author={Timothy P. Lillicrap and Jonathan J. Hunt and Alexander Pritzel and Nicolas Heess and Tom Erez and Yuval Tassa and David Silver and Daan Wierstra},
-      year={2019},
-      eprint={1509.02971},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
-}
-
-@INPROCEEDINGS{mastalli2020crocoddyl,
-  author={Mastalli, Carlos and Budhiraja, Rohan and Merkt, Wolfgang and Saurel, Guilhem and Hammoud, Bilal and Naveau, Maximilien and Carpentier, Justin and Righetti, Ludovic and Vijayakumar, Sethu and Mansard, Nicolas},
-  booktitle={2020 IEEE International Conference on Robotics and Automation (ICRA)}, 
-  title={{Crocoddyl: An Efficient and Versatile Framework for Multi-Contact Optimal Control}}, 
-  year={2020},
-  volume={},
-  number={},
-  pages={2536-2542},
-  doi={10.1109/ICRA40945.2020.9196673}}
-
-@misc{drake,
- author = {Russ Tedrake and the Drake Development Team},
- title = {{Drake: Model-Based Design and Verification for Robotics}},
- year = {2019},
- url = {https://drake.mit.edu}
-}
-
-@misc{stable-baselines3,
-  author = {Raffin, Antonin and Hill, Ashley and Ernestus, Maximilian and Gleave, Adam and 
-Kanervisto, Anssi and Dormann, Noah},
-  title = {{Stable Baselines3}},
-  year = {2019},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/DLR-RM/stable-baselines3}},
-}
-
-@misc{tensorflow2015-whitepaper,
-title={ {TensorFlow}: {Large-Scale Machine Learning on Heterogeneous Systems}},
-url={https://www.tensorflow.org/},
-note={Software available from tensorflow.org},
-author={
-    Mart\'{\i}n~Abadi and
-    Ashish~Agarwal and
-    Paul~Barham and
-    Eugene~Brevdo and
-    Zhifeng~Chen and
-    Craig~Citro and
-    Greg~S.~Corrado and
-    Andy~Davis and
-    Jeffrey~Dean and
[…]
