+++ docs/reference/debug.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Clean
-make clean
-
-# Compile raising warnings as errors
-make html SPHINXOPTS="-W"
