+++ docs/reference/source/project.py
@@ -1,23 +1,0 @@
-# -- Project information -----------------------------------------------------
-
-project = 'Torque-limited Simple Pendulum'
-author  = 'Felix Wiebe, Jonathan Babel, Shivesh Kumar, Shubham Vyas, Daniel Harnack, Melya Boukheddimi, Mihaela Popescu, Frank Kirchner'
-
-# Project code name
-codename = 'simple_pendulum'
-
-# Project code directory
-codedir = 'software/python'
-
-# Report title
-report_title = rf'Torque-limited\\Simple Pendulum:\\Reference'
-report_author = r'''
-Felix Wiebe, Jonathan Babel,\\
-\vspace*{0.5cm}
-{\fontsize{10}{10}\selectfont Contributors:}\\
-\vspace*{0.2cm}
-Shivesh Kumar, Shubham Vyas, Daniel Harnack,\\
-Melya Boukheddimi, Mihaela Popescu, Frank Kirchner'''
-
-# Logo
-logo = 'logo.jpg'
