+++ docs/_sources/simple_pendulum.controllers.tvlqr.rst.txt
@@ -1,0 +1,12 @@
+
+:autogenerated:
+
+.. automodule:: simple_pendulum.controllers.tvlqr
+    
+
+
+    .. toctree::
+       :maxdepth: 1
+
+       simple_pendulum.controllers.tvlqr.tvlqr
+
