+++ docs/_sources/simple_pendulum.utilities.performance_profiler.rst.txt
@@ -1,0 +1,10 @@
+
+:autogenerated:
+
+.. currentmodule:: performance_profiler
+
+.. automodule:: simple_pendulum.utilities.performance_profiler
+    :members: motor_send_n_commands, motor_speed_test, profiler
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
