+++ docs/_sources/simple_pendulum.reinforcement_learning.ddpg.noise.rst.txt
@@ -1,0 +1,10 @@
+
+:autogenerated:
+
+.. currentmodule:: noise
+
+.. automodule:: simple_pendulum.reinforcement_learning.ddpg.noise
+    :members: OUActionNoise
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
