+++ docs/_sources/simple_pendulum.controllers.ddpg.ddpg_controller.rst.txt
@@ -1,0 +1,10 @@
+
+:autogenerated:
+
+.. currentmodule:: ddpg_controller
+
+.. automodule:: simple_pendulum.controllers.ddpg.ddpg_controller
+    :members: ddpg_controller
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource

[…]
