### 2022-04-02 5f3e148
removed crocoddyl from setup.py, recommend apt installation of crocoddyl in installation guide
