+++ docs/reference/source/chapters/00.installation_guide.rst
@@ -174,6 +174,6 @@ This was the final installation step. Your system is now prepared to run all cod
 OPTIONAL: Crocoddyl and Gepetto Viewer
 ------------------------
 
-For installing optimal control library crocoddyl, we refer to the instructions provided in the `Crocoddyl <https://github.com/loco-3d/crocoddyl>`_ and recommend the installation through robotpkg.
+For installing the optimal control library Crocoddyl, we refer to the instructions provided in the `Crocoddyl github repository <https://github.com/loco-3d/crocoddyl>`_ and recommend the installation through robotpkg.
 
-The `Crocoddyl <https://github.com/loco-3d/crocoddyl>`_ has an interface to the gepetto-viewer for visualization. For installing the gepetto viewer we refer to their `github repository <https://github.com/Gepetto/gepetto-viewer-corba>`_.
+`Crocoddyl <https://github.com/loco-3d/crocoddyl>`_ has an interface to the gepetto-viewer for visualization. For installing the gepetto viewer we refer to their `github repository <https://github.com/Gepetto/gepetto-viewer-corba>`_.
