### 2022-04-04 dad8180
Added README for SOS RoA estimation
