+++ software/python/simple_pendulum/controllers/lqr/roa/README.md
@@ -1,4 +1,4 @@
-# Region of Attraction(RoA) estimation #
+# Region of Attraction (RoA) estimation 
 
 The RoA estimation is a process that can be applied to the closed-loop dynamics of a system, i.e. Plant + Controller, in order to analize its state-space behaviour. Two different methods have been implemented here for studying the dynamics of the underactuated pendulum coupled with an LQR controller. 
 
@@ -13,16 +13,16 @@ First, the simple pendulum plant has been put in polynomial form via Taylor appr
 
 The LQR controller has been initialized to stabilize our closed-loop system around the up-right position. From that it has been obtained a state feedback matrix K and a matrix S, which is the solution of the matrix Lyapunov equation. The last one is very usefull because it can be used to ontain the Lyapunov function as $$V(x) = x^TSx$$ 
 After obtaining the Lyapunov function the feasibility problem can be formulated as 
-$$ -\dot{V}(x) + \lambda(x)(V(x)-\rho)\ is\ SOS\quad and \quad \lambda(x)\ is\ SOS  $$
-This is coming from the so called "S-procedure" which makes use of the concept of Lagrangia multiplier to fomulate the Lyapunov conditions.  
+$ -\dot{V}(x) + \lambda(x)(V(x)-\rho)\ is\ SOS\quad and \quad \lambda(x)\ is\ SOS  $
+This is coming from the so called "S-procedure" which makes use of the concept of Lagrangian multiplier to fomulate the Lyapunov conditions.  
 It is important to highlight that other contraints are needed to include the torque inputs limits in the optimization problem.
 
 Eventually, the above problem only verify one-sublevel set of the Lyapunov function. In order to obtain the best estimation, searching for the largest rho  that can satisfy these conditions is necessary. The two different methods that have been implemented are described below.
 
-###  $\rho$  maximization: Simple line search #
+###  $\rho$  maximization: Simple line search 
 Since the problem is convex with rho fixed, and rho is just a scalar, a simple line search on rho can be performed to find the maximum for which the convex optimization returns a feasible solution. In particular, a bisection-like algorithm has been implemented here for such a purpose. However, since this formulation can be computationally heavy then other formulations might be considered.
 
-### $\rho$ maximization: Equality-constrained formulation #
+### $\rho$ maximization: Equality-constrained formulation 
 This is an important variation since it makes use of the "S-procedure" to make the problem be jointly convex in $\lambda$(x) and $\rho$, so a single convex optimization is needed.  
 Under the assumption that $\dot{V}(x)$ is negative-definite at the fixed point, the problem can be written as
 
@@ -31,8 +31,18 @@ $$ \begin{equation} \begin{aligned} \max_{\rho,\lambda} \quad & \rho\\
 
 Also in this case, input limits have then to be included to obtain the desired result.
 
+## Probabilistic Method
+For simple systems, such as the torque limited simple pendulum under LQR control, the RoA can be estimated by just evaluating the Lyapunov conditions for initial states from a continuously shrinking estimate of the ROA. This approach was introduced in:
+> E. Najafi, R. Babuška, and G. A. D. Lopes, “A fast sampling method for estimating the domain of attraction,” Nonlinear Dyn, vol. 86, no. 2, pp. 823–834, Oct. 2016, doi: 10.1007/s11071-016-2926-7.
 
-## Usage #
+The RoA is estimated by a sublevel set of a quadratic Lyapunov function:
+
[…]
