+++ software/python/simple_pendulum/controllers/lqr/roa/README.md
@@ -41,7 +41,8 @@ The RoA is estimated by a sublevel set of a quadratic Lyapunov function:
 
 Where  <img src="https://render.githubusercontent.com/render/math?math=\bar{\mathbf{x}} =\mathbf{x} - \mathbf{x}^\star"> denotes the deviation from the fixed point <img src="https://render.githubusercontent.com/render/math?math=\mathbf{x}^\star">
 
-A random initial condition <img src="https://render.githubusercontent.com/render/math?math=V(\hat{\mathbf{x}})"> is sampled from <img src="https://render.githubusercontent.com/render/math?math=\mathcal{B}"> and the Lyapunov conditions (<img src="https://render.githubusercontent.com/render/math?math=V(\hat{\mathbf{x}}) > 0$"> and <img src="https://render.githubusercontent.com/render/math?math=\dot{V}(\hat{\mathbf{x}}) = \nabla V \mathbf{f}(\mathbf{\hat{x}}) < 0">) are evaluated. If these conditions are satisfied, the next initial state is sampled. However, if these conditions are not met, <img src="https://render.githubusercontent.com/render/math?math=\rho = V(\hat{\mathbf{x}})">
+A random initial condition <img src="https://render.githubusercontent.com/render/math?math=V(\hat{\mathbf{x}})"> is sampled from <img src="https://render.githubusercontent.com/render/math?math=\mathcal{B}"> and the Lyapunov conditions (<img src="https://render.githubusercontent.com/render/math?math=V(\hat{\mathbf{x}}) > 0$"> and <img src="https://render.githubusercontent.com/render/math?math=\dot{V}(\hat{\mathbf{x}}) = \nabla V \mathbf{f}(\mathbf{\hat{x}}) < 0">) are evaluated. If these conditions are satisfied, the next initial state is sampled. However, if these conditions are not met, the estimate is shrunk, such that <img src="https://render.githubusercontent.com/render/math?math=\rho = V(\hat{\mathbf{x}})">
+
 ## Usage 
 Some usage example can be found in the [examples folder](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/tree/master/software/python/examples). The different methods can be compared in simulation with
 
