+++ software/python/simple_pendulum/controllers/lqr/roa/README.md
@@ -11,23 +11,24 @@ Furthermore, any sublevel set of a Lyapunov function is also an invariant set. T
 
 First, the simple pendulum plant has been put in polynomial form via Taylor approximation around the up-right position, which will be the goal of the LQR controller. From different tests it seems that at least the third order approximation is necessary to obtain a satisfying result.
 
-The LQR controller has been initialized to stabilize our closed-loop system around the up-right position. From that it has been obtained a state feedback matrix K and a matrix S, which is the solution of the matrix Lyapunov equation. The last one is very usefull because it can be used to ontain the Lyapunov function as $$V(x) = x^TSx$$ 
-After obtaining the Lyapunov function the feasibility problem can be formulated as 
-$ -\dot{V}(x) + \lambda(x)(V(x)-\rho)\ is\ SOS\quad and \quad \lambda(x)\ is\ SOS  $
+The LQR controller has been initialized to stabilize our closed-loop system around the up-right position. From that it has been obtained a state feedback matrix K and a matrix S, which is the solution of the matrix Lyapunov equation. The last one is very usefull because it can be used to ontain the Lyapunov function as <img src="https://render.githubusercontent.com/render/math?math=V(x) = x^TSx">  
+After obtaining the Lyapunov function the feasibility problem can be formulated as   
+<img src="https://render.githubusercontent.com/render/math?math=-\dot{V}(x) + \lambda(x)(V(x)-\rho)\ is\ SOS\quad and \quad \lambda(x)\ is\ SOS ">  
 This is coming from the so called "S-procedure" which makes use of the concept of Lagrangian multiplier to fomulate the Lyapunov conditions.  
 It is important to highlight that other contraints are needed to include the torque inputs limits in the optimization problem.
 
 Eventually, the above problem only verify one-sublevel set of the Lyapunov function. In order to obtain the best estimation, searching for the largest rho  that can satisfy these conditions is necessary. The two different methods that have been implemented are described below.
 
-###  $\rho$  maximization: Simple line search 
-Since the problem is convex with rho fixed, and rho is just a scalar, a simple line search on rho can be performed to find the maximum for which the convex optimization returns a feasible solution. In particular, a bisection-like algorithm has been implemented here for such a purpose. However, since this formulation can be computationally heavy then other formulations might be considered.
+###  <img src="https://render.githubusercontent.com/render/math?math=\rho">  maximization: Simple line search 
+Since the problem is convex with rho fixed, and rho is just a scalar, a simple line search on <img src="https://render.githubusercontent.com/render/math?math=\rho"> can be performed to find the maximum for which the convex optimization returns a feasible solution. In particular, a bisection-like algorithm has been implemented here for such a purpose. However, since this formulation can be computationally heavy then other formulations might be considered.
 
-### $\rho$ maximization: Equality-constrained formulation 
-This is an important variation since it makes use of the "S-procedure" to make the problem be jointly convex in $\lambda$(x) and $\rho$, so a single convex optimization is needed.  
-Under the assumption that $\dot{V}(x)$ is negative-definite at the fixed point, the problem can be written as
+### <img src="https://render.githubusercontent.com/render/math?math=\rho"> maximization: Equality-constrained formulation 
[…]
