### 2022-04-06 af0e78f
Improved RoA estimation README and added results plots
