+++ software/python/examples/README.md
@@ -116,7 +116,29 @@ The controller computes an initial guess for a swingup trajectory and then swing
 
 Requirements: Drake (optional)
 
-If Drake is installed the symbolic libary of Drake will be used, else sympy is used.
+If Drake is installed the symbolic libary of Drake will be used, else sympy is 
+used.
+
+## Region of Attraction estimation
+
+The Region of Atttraction estimation procedure is described [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/software/python/simple_pendulum/controllers/lqr/roa).   
+The pendulum parameters are fixed to mass = 0.57288, length = 0.5, 
+damping = 0.15, gravity = 9.81, coulomb_fric = 0.0. Also the goal position is fixed to the up-right position.
+
+The different used methods can be compared in simulation with
+
+    python plot_roa_estimations.py
+Here the torque limits can be modified to see different results, they should be at most five different. The choosed torque limits are 0.1, 0.5, 1, 2, 3.
+
+The RoA certification reliability can be verified with
+
+    python verify_roa_estimation.py
+where 500 different initial conditions have been chosen to verify the theory. Choosing a lot of different initial conditions will slow down the execution. The torque limit has been fixed to 4, but it is not a restrictive value for the code functionality.
+
+Furthermore, the effects of the Taylor approximation on the closed-loop dynamics can be seen with
+
+    python taylorApprox_roa_sos.py
+where the maximum approximation order has been fixed to 7, enough to see the interesting results. The torque limit has been fixed to 8, the high value is due to the need of analyzing the all range of angles fro 0 to pi.
 
 ## Analysis
 
@@ -127,6 +149,8 @@ The policies of the energy shaping, LQR, SAC and DDPG controllers can be plotted
     python plot_controller.py
 
 The controller to be plotted can be specified in the script.
+
+The RoA estimation plots can be found in the result folder [here](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/results/RoA_estimation_plots).
 
 ### Benchmarking
 
