+++ docs/_sources/chapters/08.policy_based_controllers/83.lqr.rst.txt
@@ -80,3 +80,4 @@ Comments
 
 Without torque limits the LQR controller can drive the pendulum up from any position in a straight way. In practice this controller should only be used for stabilizing the pendulum at the unstable fixpoint. If the controller would require a torque larger than the pendulums torque limit, the controller returns None instead. This makes it possible to combine this controller with another controller and only use the LQR control if the output is not None. The region of attraction where this controller is able to stabilize the pendulum depends on the pendulum parameters and especially its torque limits.
 
+.. include:: 83.roa.rst
