### 2022-04-07 1bec167
LQR documentation refactored
