### 2022-04-07 89c2df0
LQR ROA documentation syntax errors removed
