### 2022-04-07 a66921c
Removed redundant ROA doc inclusion
