### 2022-04-13 #3884 @shivesh1210
@gkthiruvathukal Is there any update on the status of our paper? It has been six months already since we submitted our work. It would be great if you could assist in speeding up the review process. In the meantime, we have also created a documentation website: https://dfki-ric-underactuated-lab.github.io/torque_limited_simple_pendulum/ and setup a CI pipeline for our work.
