+++ software/python/simple_pendulum/controllers/ilqr/__init__.py
@@ -1,4 +1,0 @@
-"""
-iLQR Model Predictive Control
-=============================
-"""
