+++ software/python/simple_pendulum/controllers/energy_shaping/__init__.py
@@ -1,4 +1,0 @@
-"""
-Energy Shaping Control
-======================
-"""
