+++ software/python/simple_pendulum/controllers/tvlqr/__init__.py
@@ -1,4 +1,0 @@
-"""
-Time-varying Linear Quadratic Regulator (TVLQR)
-===============================================
-"""

[…]
