+++ .github/workflows/docs.yml
@@ -185,4 +185,4 @@ jobs:
         git config --local user.name "github-actions[bot]"
         git pull
         git add $pdfdir/${{ github.event.repository.name }}.pdf && git commit --allow-empty -m "Updated Report"
-        git push -u
+        git push -u origin ${GITHUB_REF##*/}
