+++ .github/workflows/pytest.yml
@@ -31,6 +31,11 @@ jobs:
       uses: actions/checkout@v3
       with:
         fetch-depth: 0
+    - name: TEST
+      run: |
+        echo ${GITHUB_REF##*/}
+        echo =================
+        echo $GITHUB_REF
     # Load environment variables from config.env =================
     - name: Load config variables into env
       run: |
@@ -130,7 +135,7 @@ jobs:
         git config --local user.name "github-actions[bot]"
         git pull
         git add cov/ub18cov.svg && git commit --allow-empty -m "Badge updated"
-        git push -u origin ${GITHUB_REF##*/}
+        git push -u origin master
     - name: Close
       if: steps.tests.outcome != 'success'
       run: |
@@ -250,7 +255,7 @@ jobs:
         git config --local user.name "github-actions[bot]"
         git pull
         git add cov/ub20cov.svg && git commit --allow-empty -m "Badge updated"
-        git push -u origin ${GITHUB_REF##*/}
+        git push -u origin master
     - name: Close
       if: steps.tests.outcome != 'success'
       run: |
