+++ .github/workflows/pytest.yml
@@ -31,11 +31,6 @@ jobs:
       uses: actions/checkout@v3
       with:
         fetch-depth: 0
-    - name: TEST
-      run: |
-        echo ${GITHUB_REF##*/}
-        echo =================
-        echo $GITHUB_REF
     # Load environment variables from config.env =================
     - name: Load config variables into env
       run: |
