### 2022-04-27 #3884 @fwiebe
Hi @gkthiruvathukal ,
- we created a tagged release of the software (tag: v1.0.0) (https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/releases/tag/v1.0.0)
- The software is archived on Zenodo (https://zenodo.org/record/6498036), The Zenodo DOI is: 10.5281/zenodo.6498036
