### 2022-05-09 b9d9f2f
cleaned up bibtex file
