+++ docs/paper/paper.bib
@@ -1,9 +1,9 @@
 @article{weiwei2004iterative,
-  	author = {{Weiwei}, L.. and {Todorov}, E.},
-  	Journal = {International Conference on Informatics in Control, Automation and Robotics},
-  	Title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
-  	Year = 2004,
-    pages = 222-229
+  author = {{Weiwei}, Li and {Todorov}, Emanuel},
+  journal = {International Conference on Informatics in Control, Automation and Robotics},
+  title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
+  year = {2004},
+  pages = {222-229}
 }
 
 @article{hargraves1987direct,
@@ -17,28 +17,27 @@
 }
 
 @misc{haarnoja2018soft,
-      title={{Soft Actor-Critic: Off-Policy Maximum Entropy Deep Reinforcement Learning with a 
-Stochastic Actor}}, 
-      author={Tuomas Haarnoja and Aurick Zhou and Pieter Abbeel and Sergey Levine},
-      year={2018},
-      eprint={1801.01290},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
+  title={{Soft Actor-Critic: Off-Policy Maximum Entropy Deep Reinforcement Learning with a Stochastic Actor}},
+  author={Tuomas Haarnoja and Aurick Zhou and Pieter Abbeel and Sergey Levine},
+  year={2018},
+  eprint={1801.01290},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG}
 }
 
 @misc{lillicrap2019continuous,
-      title={{Continuous Control with Deep Reinforcement Learning}}, 
-      author={Timothy P. Lillicrap and Jonathan J. Hunt and Alexander Pritzel and Nicolas Heess and Tom Erez and Yuval Tassa and David Silver and Daan Wierstra},
-      year={2019},
-      eprint={1509.02971},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
+  title={{Continuous Control with Deep Reinforcement Learning}},
+  author={Timothy P. Lillicrap and Jonathan J. Hunt and Alexander Pritzel and Nicolas Heess and Tom Erez and Yuval Tassa and David Silver and Daan Wierstra},
+  year={2019},
+  eprint={1509.02971},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG}
 }
 
 @INPROCEEDINGS{mastalli2020crocoddyl,
   author={Mastalli, Carlos and Budhiraja, Rohan and Merkt, Wolfgang and Saurel, Guilhem and Hammoud, Bilal and Naveau, Maximilien and Carpentier, Justin and Righetti, Ludovic and Vijayakumar, Sethu and Mansard, Nicolas},
-  booktitle={2020 IEEE International Conference on Robotics and Automation (ICRA)}, 
-  title={{Crocoddyl: An Efficient and Versatile Framework for Multi-Contact Optimal Control}}, 
+  booktitle={2020 IEEE International Conference on Robotics and Automation (ICRA)},
+  title={{Crocoddyl: An Efficient and Versatile Framework for Multi-Contact Optimal Control}},
   year={2020},
   volume={},
   number={},
@@ -46,15 +45,14 @@ Stochastic Actor}},
   doi={10.1109/ICRA40945.2020.9196673}}
 
 @misc{drake,
- author = {Russ Tedrake and the Drake Development Team},
[…]
