+++ docs/paper/paper.bib
@@ -1,11 +1,3 @@
-@article{weiwei2004iterative,
-  author = {{Weiwei}, Li and {Todorov}, Emanuel},
-  journal = {International Conference on Informatics in Control, Automation and Robotics},
-  title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
-  year = {2004},
-  pages = {222-229}
-}
-
 @article{hargraves1987direct,
   title={{Direct Trajectory Optimization Using Nonlinear Programming and Collocation}},
   author={Hargraves, Charles R and Paris, Stephen W},
@@ -14,6 +6,14 @@
   number={4},
   pages={338--342},
   year={1987}
+}
+
+@article{weiwei2004iterative,
+  author = {{Weiwei}, Li and {Todorov}, Emanuel},
+  journal = {International Conference on Informatics in Control, Automation and Robotics},
+  title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
+  year = {2004},
+  pages = {222-229}
 }
 
 @misc{haarnoja2018soft,
