### 2022-05-09 31002b9
added DOIs when existing for bibtex entries
