+++ docs/paper/paper.bib
@@ -5,7 +5,8 @@
   volume={10},
   number={4},
   pages={338--342},
-  year={1987}
+  year={1987},
+  doi={10.2514/6.1986-2000}
 }
 
 @article{weiwei2004iterative,
@@ -13,7 +14,8 @@
   journal = {International Conference on Informatics in Control, Automation and Robotics},
   title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
   year = {2004},
-  pages = {222-229}
+  pages = {222-229},
+  doi={{10.5220/0001143902220229}}
 }
 
 @misc{haarnoja2018soft,
@@ -22,7 +24,8 @@
   year={2018},
   eprint={1801.01290},
   archivePrefix={arXiv},
-  primaryClass={cs.LG}
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.1801.01290}
 }
 
 @misc{lillicrap2019continuous,
@@ -31,7 +34,8 @@
   year={2019},
   eprint={1509.02971},
   archivePrefix={arXiv},
-  primaryClass={cs.LG}
+  primaryClass={cs.LG},
+  doi={10.48550/arXiv.1509.02971}
 }
 
 @INPROCEEDINGS{mastalli2020crocoddyl,
@@ -48,65 +52,74 @@
   author = {Russ Tedrake and {the Drake Development Team}},
   title = {{Drake: Model-Based Design and Verification for Robotics}},
   year = {2019},
-  url = {https://drake.mit.edu}
+  url = {https://drake.mit.edu},
+  doi = {}
 }
 
-@misc{stable-baselines3,
-  author = {Raffin, Antonin and Hill, Ashley and Ernestus, Maximilian and Gleave, Adam and Kanervisto, Anssi and Dormann, Noah},
-  title = {{Stable Baselines3}},
-  year = {2019},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/DLR-RM/stable-baselines3}},
+@article{stable-baselines3,
+  author  = {Antonin Raffin and Ashley Hill and Adam Gleave and Anssi Kanervisto and Maximilian Ernestus and Noah Dormann},
+  title   = {Stable-Baselines3: Reliable Reinforcement Learning Implementations},
+  journal = {Journal of Machine Learning Research},
+  year    = {2021},
+  volume  = {22},
+  number  = {268},
+  pages   = {1-8},
+  url     = {http://jmlr.org/papers/v22/20-1364.html},
+  doi     = {}
 }
 
 @misc{tensorflow2015-whitepaper,
-  title={ {TensorFlow}: {Large-Scale Machine Learning on Heterogeneous Systems}},
-  url={https://www.tensorflow.org/},
-  note={Software available from tensorflow.org},
-  author={
-      Mart\'{\i}n~Abadi and
-      Ashish~Agarwal and
-      Paul~Barham and
-      Eugene~Brevdo and
-      Zhifeng~Chen and
-      Craig~Citro and
-      Greg~S.~Corrado and
-      Andy~Davis and
-      Jeffrey~Dean and
-      Matthieu~Devin and
-      Sanjay~Ghemawat and
-      Ian~Goodfellow and
-      Andrew~Harp and
-      Geoffrey~Irving and
-      Michael~Isard and
-      Yangqing Jia and
-      Rafal~Jozefowicz and
-      Lukasz~Kaiser and
-      Manjunath~Kudlur and
-      Josh~Levenberg and
-      Dandelion~Man\'{e} and
-      Rajat~Monga and
-      Sherry~Moore and
-      Derek~Murray and
-      Chris~Olah and
[…]
