### 2022-05-09 #3884 @fwiebe
Hi,
as the pdf compiles successfully, I think the error arises from the bot checking the references for missing DOIs. I added DOIs for the references where they exist and added empty {} for the 3 BibTeX entries where no DOI exists. I hope this fixes the problem.
Do we need to create a new tagged release for checking if this resolves the issue?
