+++ docs/paper/paper.bib
@@ -15,7 +15,7 @@
   title = {{Iterative Linear Quadratic Regulator Design for Nonlinear Biological Movement Systems.}},
   year = {2004},
   pages = {222-229},
-  doi={{10.5220/0001143902220229}}
+  doi={10.5220/0001143902220229}
 }
 
 @misc{haarnoja2018soft,
