### 2022-05-27 28cae49
Added probabilistic time-varying RoA estimation
