+++ software/python/simple_pendulum/controllers/tvlqr/roa/plot.py
@@ -1,0 +1,178 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import patches
+import matplotlib as mpl
+import mpl_toolkits.mplot3d.art3d as art3d
+mpl.use('WebAgg')
+from scipy.spatial import ConvexHull, convex_hull_plot_2d
+from matplotlib.collections import LineCollection
+
+from simple_pendulum.controllers.tvlqr.roa.utils import projectedEllipseFromCostToGo, getEllipseContour
+
+def get_ellipse_params(rho,M):
+    """
+    Returns ellipse params (excl center point)
+    """
+
+    #eigenvalue decomposition to get the axes
+    w,v=np.linalg.eigh(M/rho) 
+
+    try:
+        #let the smaller eigenvalue define the width (major axis*2!)
+        width = 2/float(np.sqrt(w[0]))
+        height = 2/float(np.sqrt(w[1]))
+        
+        #the angle of the ellipse is defined by the eigenvector assigned to the smallest eigenvalue (because this defines the major axis (width of the ellipse))
+        angle = np.rad2deg(np.arctan2(v[:,0][1],v[:,0][0]))
+
+    except:
+        print("paramters do not represent an ellipse.")
+
+    return width,height,angle
+
+def get_ellipse_patch(px,py,rho,M,alpha_val=1,linec="red",facec="none",linest="solid"):
+    """
+    return an ellipse patch
+    """
+    w,h,a = get_ellipse_params(rho,M)
+    return patches.Ellipse((px,py), w, h, a, alpha=alpha_val,ec=linec,facecolor=facec,linestyle=linest)
+
+def plot_ellipse(px,py,rho, M, save_to=None, show=True):
+    p=get_ellipse_patch(px,py,rho,M)
+    
+    fig, ax = plt.subplots()
+    ax.add_patch(p)
+    l=np.max([p.width,p.height])
+
+    ax.set_xlim(px-l/2,px+l/2)
+    ax.set_ylim(py-l/2,py+l/2)
+
+    ax.grid(True)
+
+    if not (save_to is None):
+        plt.savefig(save_to)
+    if show:
+        plt.show()
+
+def plotFirstLastEllipses(rho, x0, goal, x0_traj, S_t, time):
+    fig,ax = plt.subplots(1,2, figsize=(18,8))
+    p0 = get_ellipse_patch(np.array(x0_traj).T[0][0],np.array(x0_traj).T[0][1],rho[0],S_t.value(time[0]),linec= "black")
+    ax[0].scatter([x0[0]],[x0[1]],color="black",marker="o")
+    ax[0].add_patch(p0)
+    ax[0].grid(True)
+    ax[0].set_xlabel("x")
+    ax[0].set_ylabel(r"$\dot{x}$")
+    ax[0].title.set_text('First ellipse')
+    pl = get_ellipse_patch(np.array(x0_traj).T[-1][0],np.array(x0_traj).T[-1][1],rho[-1],S_t.value(time[-1]),linec= "black")
+    ax[1].scatter([goal[0]],[goal[1]],color="black",marker="x")
+    ax[1].add_patch(pl)
+    ax[1].grid(True)
+    ax[1].set_xlabel("x")
+    ax[1].set_ylabel(r"$\dot{x}$")
+    ax[1].title.set_text('Last ellipse')
+
+def plotRhoEvolution(rho, x0_traj, time, N):
+    fig = plt.figure()
+    plt.title("rho evolution")
+    ax = fig.add_subplot()
+    ax.plot(np.arange(N),rho,color="yellow", label = "final rho")
+    ax2=ax.twinx()
+    d = round(len(time)/N)
[…]
