### 2022-05-31 6a3da84
fix pytest ubuntu 2020 workflow
