+++ docs/_sources/simple_pendulum.controllers.tvlqr.roa.rst.txt
@@ -1,0 +1,14 @@
+
+:autogenerated:
+
+.. automodule:: simple_pendulum.controllers.tvlqr.roa
+    
+
+
+    .. toctree::
+       :maxdepth: 1
+
+       simple_pendulum.controllers.tvlqr.roa.plot
+       simple_pendulum.controllers.tvlqr.roa.probabilistic
+       simple_pendulum.controllers.tvlqr.roa.utils
+
