+++ docs/_sources/simple_pendulum.controllers.tvlqr.roa.utils.rst.txt
@@ -1,0 +1,10 @@
+
+:autogenerated:
+
+.. currentmodule:: utils
+
+.. automodule:: simple_pendulum.controllers.tvlqr.roa.utils
+    :members: direct_sphere, getEllipseContour, projectedEllipseFromCostToGo, quad_form, sample_from_ellipsoid
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
