+++ docs/_sources/simple_pendulum.controllers.tvlqr.roa.plot.rst.txt
@@ -1,0 +1,10 @@
+
+:autogenerated:
+
+.. currentmodule:: plot
+
+.. automodule:: simple_pendulum.controllers.tvlqr.roa.plot
+    :members: get_ellipse_params, get_ellipse_patch, plotFirstLastEllipses, plotFunnel, plotFunnel3d, plotRhoEvolution, plot_ellipse
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
