### 2022-05-31 8e78043
update of drake api in direct collocation solver
