### 2022-06-01 #3884 @xuanxu
The origin of those errors is in the [code for the figures](https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/blob/master/docs/paper/paper.md?plain=1#L95) in the paper.
The `#id .class` is not valid. I suggest that is removed from every figure.
