### 2022-06-01 9723c18
removed #id .class in paper figures
