### 2022-06-01 #3884 @danielskatz
@shivesh1210 - I'm the AEiC on duty this week who will handle the rest of the publication process.  I have a fairly large number of suggestion for the paper (mostly style and consistent language).  Please merge https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/pull/22 or let me know what you disagree with, then we can continue the process.
