### 2022-06-01 #3884 @danielskatz
@shivesh1210 - I found a couple more small things in my hopefully final check: https://github.com/dfki-ric-underactuated-lab/torque_limited_simple_pendulum/pull/23
